test: Created the three employee test cases
Three test cases were created for the employee microservice, which are:
employeeCRUD: To test the creation, update, deletion and loading of
employee data.
findAllByOrderByBirthDateAsc: To verify the data, they are listed ordered
by date of birth.
findAllByOrderBySalaryDesc: To verify the data, it is listed ordered by
salary and correctly calculates the AFP amount.